Chunks: Fixed chunks being the same in all cords.
This problem occurs due to my lack of understanding in C++, an object and a pointer is not the same, while both of them has its pro and cons, an object does prevents a memory leak, BUT it will going to rewrite the whole paletted block array, since it is being passed by reference to the Chunks, resulting all the chunks being the same data. However, copying the object into another object would give a proper result. And as for memory management? Well, I assume it will be destroyed after the Chunk were deleted.
Properly copy BiomeArray into chunk for proper biome results.