Fixed biome id error on chunk axis x=0, z=0
The root cause of the problem is unjustified, but the changes made here are to make sure that SimpleChunkManager destructed using the C++ class destructors rather than destroying objects in the function itself. This does fix the biome to generate its values properly (The internals biome generator works flawlessly without any issues). And of course, "If it works, it works".